chore: rename component helpers to util
- rename component helper files from -helpers to .util
