+++ src/Frontend/Components/AggregatedAttributionsPanel/AccordionPanel.util.ts
