+++ src/Frontend/Components/AttributionColumn/LicenseSubPanel.tsx
@@ -22,7 +22,7 @@ import { useAppDispatch, useAppSelector } from '../../state/hooks';
 import { getFrequentLicensesNameOrder } from '../../state/selectors/all-views-resource-selectors';
 import { Confirm } from '../ConfirmationDialog/ConfirmationDialog';
 import { TextBox } from '../InputElements/TextBox';
-import { getLicenseTextLabelText } from './attribution-column-helpers';
+import { getLicenseTextLabelText } from './AttributionColumn.util';
 import { PackageAutocomplete } from './PackageAutocomplete';
 import { attributionColumnClasses } from './shared-attribution-column-styles';
