+++ src/Frontend/Components/ProgressBar/ProgressBar.util.tsx
