+++ src/Frontend/Components/ProjectStatisticsPopup/ProjectStatisticsPopup.util.ts
