+++ src/Frontend/Components/ResourcePathPopup/ResourcePathPopup.util.ts
