+++ src/Frontend/Components/ResourcePathPopup/__tests__/ResourcePathPopup.util.test.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 import { AttributionsToResources } from '../../../../shared/shared-types';
-import { getAllResourcePathsForAttributions } from '../resource-path-popup-helpers';
+import { getAllResourcePathsForAttributions } from '../ResourcePathPopup.util';
 
 describe('getAllResourcePathsForAttributions', () => {
   it('gets resource paths for attributions', () => {
